+++ packages/assets-controllers/src/MultichainBalancesController/MultichainBalancesController.test.ts
@@ -7,7 +7,6 @@ import {
   EthMethod,
   BtcScope,
   EthScope,
-  SolScope,
 } from '@metamask/keyring-api';
 import { KeyringTypes } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';