+++ packages/assets-controllers/src/MultichainBalancesController/MultichainBalancesController.ts
@@ -137,7 +137,7 @@ export class MultichainBalancesController extends BaseController<
   MultichainBalancesControllerState,
   MultichainBalancesControllerMessenger
 > {
-  #tracker: BalancesTracker;
+  readonly #tracker: BalancesTracker;
 
   constructor({
     messenger,