+++ packages/assets-controllers/src/index.ts
@@ -166,3 +166,17 @@ export type {
   MultichainBalancesControllerEvents,
   MultichainBalancesControllerMessenger,
 } from './MultichainBalancesController';
+
+export {
+  MultichainAssetsController,
+  getDefaultMultichainAssetsControllerState,
+} from './MultichainAssetsController';
+
+export type {
+  MultichainAssetsControllerState,
+  MultichainAssetsControllerGetStateAction,
+  MultichainAssetsControllerStateChangeEvent,
+  MultichainAssetsControllerActions,
+  MultichainAssetsControllerEvents,
+  MultichainAssetsControllerMessenger,
+} from './MultichainAssetsController';