+++ packages/profile-sync-controller/src/sdk/authentication-jwt-bearer/services.ts
@@ -1,3 +1,5 @@
+import type { AccessToken, ErrorMessage, UserProfile } from './types';
+import { AuthType } from './types';
 import type { Env, Platform } from '../../shared/env';
 import { getEnvUrls, getOidcClientId } from '../../shared/env';
 import {
@@ -6,8 +8,6 @@ import {
   SignInError,
   ValidationError,
 } from '../errors';
-import type { AccessToken, ErrorMessage, UserProfile } from './types';
-import { AuthType } from './types';
 
 export const NONCE_URL = (env: Env) =>
   `${getEnvUrls(env).authApiUrl}/api/v2/nonce`;
@@ -47,13 +47,13 @@ type NonceResponse = {
 type PairRequest = {
   signature: string;
   // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/naming-convention
+
   raw_message: string;
   // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/naming-convention
+
   encrypted_storage_key: string;
   // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/naming-convention
+
   identifier_type: 'SIWE' | 'SRP';
 };
 
@@ -168,7 +168,7 @@ export async function authorizeOIDC(
     if (!response.ok) {
       const responseBody = (await response.json()) as {
         // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-        // eslint-disable-next-line @typescript-eslint/naming-convention
+
         error_description: string;
         error: string;
       };
@@ -222,7 +222,7 @@ export async function authenticate(
       body: JSON.stringify({
         signature,
         // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-        // eslint-disable-next-line @typescript-eslint/naming-convention
+
         raw_message: rawMessage,
       }),
     });