+++ packages/transaction-controller/src/TransactionController.ts
@@ -11,7 +11,7 @@ import type {
 import type {
   ControllerGetStateAction,
   ControllerStateChangeEvent,
-  RestrictedControllerMessenger,
+  RestrictedMessenger,
 } from '@metamask/base-controller';
 import { BaseController } from '@metamask/base-controller';
 import {
@@ -539,7 +539,7 @@ export type TransactionControllerEvents =
 /**
  * The messenger of the {@link TransactionController}.
  */
-export type TransactionControllerMessenger = RestrictedControllerMessenger<
+export type TransactionControllerMessenger = RestrictedMessenger<
   typeof controllerName,
   TransactionControllerActions | AllowedActions,
   TransactionControllerEvents | AllowedEvents,
@@ -3320,10 +3320,12 @@ export class TransactionController extends BaseController<
     provider,
     blockTracker,
     chainId,
+    networkClientId,
   }: {
     provider: Provider;
     blockTracker: BlockTracker;
     chainId: Hex;
+    networkClientId: NetworkClientId;
   }): PendingTransactionTracker {
     const ethQuery = new EthQuery(provider);
 
@@ -3331,6 +3333,7 @@ export class TransactionController extends BaseController<
       blockTracker,
       getChainId: () => chainId,
       getEthQuery: () => ethQuery,
+      getNetworkClientId: () => networkClientId,
       getTransactions: () => this.state.transactions,
       isResubmitEnabled: this.#pendingTransactionOptions.isResubmitEnabled,
       getGlobalLock: () =>